refactor : 입출력 클래스 생성
